# backend/docs/swagger.yaml
basePath: /
definitions:
  codepulse.ActionCheckReq:
    properties:
      action:
        type: string
      campaign_id:
        type: integer
      milestone_index:
        type: integer
      params: {}
      proposal_id:
        type: integer
      wallet:
        type: string
    required:
    - action
    - wallet
    type: object
  codepulse.ActionCheckResp:
    properties:
      allowed:
        type: boolean
      current_state:
        type: string
      reason_code:
        type: string
      reason_message:
        type: string
      required_role:
        type: string
      revert_error_args: {}
      revert_error_name:
        type: string
    type: object
...
info:
  contact: {}
  description: 'REST API: health, chain status, counter contract calls, bank ledger
    (PostgreSQL indexer + optional The Graph subgraph).'
  title: go-chain API
  version: 0.1.0
paths:
...
  /api/code-pulse/actions/check:
    post:
      consumes:
      - application/json
      parameters:
      - description: Action check request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codepulse.ActionCheckReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codepulse.ActionCheckResp'
      summary: Action pre-check
      tags:
      - code-pulse
...
  /api/code-pulse/admin/platform-funds:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Platform funds overview
      tags:
      - code-pulse-admin
  /api/code-pulse/admin/proposal-initiators:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List proposal initiators
      tags:
      - code-pulse-admin
    post:
      consumes:
      - application/json
      parameters:
      - description: '{ address: string }'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Add proposal initiator
      tags:
      - code-pulse-admin
  /api/code-pulse/admin/proposal-initiators/{address}:
    delete:
      parameters:
      - description: Wallet address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Remove proposal initiator
      tags:
      - code-pulse-admin
...
swagger: "2.0"